+++ src/main/java/com/pearson/statsagg/utilities/Threads.java
@@ -124,7 +124,7 @@ public class Threads {
         }
     }
     
-    public static void threadExecutorFixedPool(List threads, int nThreadPoolSize, int timeoutTime, TimeUnit timeoutTimeunit) {
+    public static void threadExecutorFixedPool(List threads, int nThreadPoolSize, long timeoutTime, TimeUnit timeoutTimeunit) {
         
         if ((threads == null) || threads.isEmpty() ||(nThreadPoolSize <= 0) || (timeoutTime <= 0) || (timeoutTimeunit == null)) {
             return;
@@ -161,7 +161,7 @@ public class Threads {
 
     }
     
-    public static void threadExecutorCachedPool(List threads, int timeoutTime, TimeUnit timeoutTimeunit) {
+    public static void threadExecutorCachedPool(List threads, long timeoutTime, TimeUnit timeoutTimeunit) {
         
         if ((threads == null) || threads.isEmpty() || (timeoutTime <= 0) || (timeoutTimeunit == null)) {
             return;
@@ -199,7 +199,7 @@ public class Threads {
 
     }
     
-    public static void shutdownThreadExecutor(ExecutorService threadExecutor, Integer timeoutTime, TimeUnit timeoutTimeunit, 
+    public static void shutdownThreadExecutor(ExecutorService threadExecutor, Long timeoutTime, TimeUnit timeoutTimeunit, 
             boolean forceShutdownIfTimeoutReached, boolean waitForeverForForceShutdownToFinish) {
 
         if (threadExecutor == null) {
